+++ packages/in_app_purchase/test/store_kit_wrappers/sk_methodchannel_apis_test.dart
@@ -85,7 +85,7 @@ void main() {
 
     test('pendingTransactions should return a valid list of transactions',
         () async {
-      expect(await SKPaymentQueueWrapper().pendingTransactions, isNotEmpty);
+      expect(await SKPaymentQueueWrapper().transactions, isNotEmpty);
     });
 
     test(